Move menu contents into a seperate component, DashMenu.js
